+++ h3ron-h3-sys/update-prebuild-bindings.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cargo clean
+cargo build --release --features "bindgen"
+cp `find  ../target/ -name 'bindings.rs'` src/prebuild_bindings.rs